ui/requirements.txt
numpy==1.23.5
Pillow==10.2.0
Requests==2.31.0
streamlit==1.30.0
